requirements.txt
attrs==24.3.0
Faker~=37.3.0
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest~=8.3.5
selenium~=4.33.0
openpyxl==3.1.2
allure-python-commons~=2.14.2
urllib3~=2.4.0